requirements/local.txt
-r base.txt

# Development Tools
django-debug-toolbar>=4.2.0,<5.0
ipython>=8.18.0,<9.0
pytest-django>=4.7.0,<5.0
black>=23.12.0,<24.0
flake8>=6.1.0,<7.0
isort>=5.13.0,<6.0 